requirements.txt
slack-bolt==1.18.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
boto3==1.34.0
requests==2.31.0 